+++ Collect/MOD10/DataAccess.py
@@ -120,30 +120,36 @@ def RetrieveData(Date, args):
     # Argument
     [output_folder, TilesVertical, TilesHorizontal,lonlim, latlim, hdf_library] = args
 
-    # Collect the data from the MODIS webpage and returns the data and lat and long in meters of those tiles
-    try:
-        Collect_data(TilesHorizontal, TilesVertical, Date, output_folder, hdf_library)
-    except:
-        print("Was not able to download the file")
-
-    # Define the output name of the collect data function
-    name_collect = os.path.join(output_folder, 'Merged.tif')
-
-    # Reproject the MODIS product to epsg_to
-    epsg_to ='4326'
-    name_reprojected = RC.reproject_MODIS(name_collect, epsg_to)
-
-    # Clip the data to the users extend
-    data, geo = RC.clip_data(name_reprojected, latlim, lonlim)
-
-    # Save the file as tiff
     FPARfileName = os.path.join(output_folder, 'SnowFrac_MOD10_unitless_8-daily_'  + Date.strftime('%Y') + '.' + Date.strftime('%m') + '.' + Date.strftime('%d') + '.tif')
-    DC.Save_as_tiff(name=FPARfileName, data=data, geo=geo, projection='WGS84')
+ 
+    if not os.path.exists(FPARfileName):
 
-    # remove the side products
-    os.remove(os.path.join(output_folder, name_collect))
-    os.remove(os.path.join(output_folder, name_reprojected))
-
+        # Collect the data from the MODIS webpage and returns the data and lat and long in meters of those tiles
+        try:
+            Collect_data(TilesHorizontal, TilesVertical, Date, output_folder, hdf_library)
+        except:
+            print("Was not able to download the file")
+        try:
+            # Define the output name of the collect data function
+            name_collect = os.path.join(output_folder, 'Merged.tif')
+        
+            # Reproject the MODIS product to epsg_to
+            epsg_to ='4326'
+            name_reprojected = RC.reproject_MODIS(name_collect, epsg_to)
+        
+            # Clip the data to the users extend
+            data, geo = RC.clip_data(name_reprojected, latlim, lonlim)
+        
+            # Save the file as tiff
+            DC.Save_as_tiff(name=FPARfileName, data=data, geo=geo, projection='WGS84')
+        
+            # remove the side products
+            os.remove(os.path.join(output_folder, name_collect))
+            os.remove(os.path.join(output_folder, name_reprojected))
+                
+        except:
+            print("Failed for date: %s" %Date)
+        
     return True
 
 def Make_TimeStamps(Startdate,Enddate):
